changed the room-constructor to contain a boolean that tells if the database should be updated or not when you create a room bc of the room-search that crates a new arraylist of rooms. BookClassForSemester is now up and running and working as far as i know. Wrote down a few notes in the grouplog to remind me/us what needs to be done. Also moved roomList to OldCode-package.
