+++ src/OldCode/RoomList.java
@@ -1,6 +1,8 @@
-package RooMe;
+package OldCode;
 
 import java.util.ArrayList;
+
+import RooMe.Room;
 
 public class RoomList {
